refactor: TopAppBar 재사용을 위해 분리
